port improvements to test container from upstream pxr

+++ tests/ci/setup/setup_php.sh
@@ -2,9 +2,15 @@
 
 # Has to be run as admin
 
+# @todo make it optional to install xdebug. It is fe. missing in sury's ppa for Xenial
+# @todo make it optional to install fpm
 # @todo make it optional to disable xdebug ?
 
 set -e
+
+echo "Installing PHP version '${1}'..."
+
+SCRIPT_DIR="$(dirname -- "$(readlink -f "$0")")"
 
 configure_php_ini() {
     # note: these settings are not required for cli config
@@ -12,7 +18,12 @@ configure_php_ini() {
     echo "always_populate_raw_post_data = -1" >> "${1}"
 
     # we disable xdebug for speed for both cli and web mode
-    phpdismod xdebug
+    # @todo make this optional
+    if which phpdismod >/dev/null 2>/dev/null; then
+        phpdismod xdebug
+    elif [ -f /usr/local/php/$PHP_VERSION/etc/conf.d/20-xdebug.ini ]; then
+        mv /usr/local/php/$PHP_VERSION/etc/conf.d/20-xdebug.ini /usr/local/php/$PHP_VERSION/etc/conf.d/20-xdebug.ini.bak
+    fi
 }
 
 # install php
@@ -26,7 +37,11 @@ if [ -z "${DEBIAN_VERSION}" ]; then
     DEBIAN_VERSION=$(cat /etc/os-release | grep 'VERSION=' | grep 'VERSION=' | sed 's/VERSION=//' | sed 's/"[0-9.]\+ *(\?//' | sed 's/)\?"//' | tr '[:upper:]' '[:lower:]' | sed 's/lts, *//' | sed 's/ \+tahr//')
 fi
 
+# @todo use native packages if requested for a specific version and that is the same as available in the os repos
+
 if [ "${PHP_VERSION}" = default ]; then
+    echo "Using native PHP packages..."
+
     if [ "${DEBIAN_VERSION}" = jessie -o "${DEBIAN_VERSION}" = precise -o "${DEBIAN_VERSION}" = trusty ]; then
         PHPSUFFIX=5
     else
@@ -55,6 +70,17 @@ if [ "${PHP_VERSION}" = default ]; then
         :
     fi
 else
+    # on GHA runners ubuntu version, php 7.4 and 8.0 seem to be preinstalled. Remove them if found
+    for PHP_CURRENT in $(dpkg -l | grep -E 'php.+-common' | awk '{print $2}'); do
+        if [ "${PHP_CURRENT}" != "php${PHP_VERSION}-common" ]; then
+            apt-get purge -y "${PHP_CURRENT}"
+        fi
+    done
+
+    # @todo use php from shivammathur/php5-ubuntu for php versions 5.4, 5.5 (see the phpxmlrpc repo)
+
+    echo "Using PHP packages from ondrej/php..."
+
     DEBIAN_FRONTEND=noninteractive apt-get install -y language-pack-en-base software-properties-common
     LC_ALL=en_US.UTF-8 add-apt-repository ppa:ondrej/php
     apt-get update
@@ -95,16 +121,40 @@ fi
 
 PHPVER=$(php -r 'echo implode(".",array_slice(explode(".",PHP_VERSION),0,2));' 2>/dev/null)
 
-configure_php_ini /etc/php/${PHPVER}/fpm/php.ini
+service "php${PHPVER}-fpm" stop || true
 
-# use a nice name for the php-fpm service, so that it does not depend on php version running
-service "php${PHPVER}-fpm" stop
-ln -s "/etc/init.d/php${PHPVER}-fpm" /etc/init.d/php-fpm
+if [ -d /etc/php/${PHPVER}/fpm ]; then
+    configure_php_ini /etc/php/${PHPVER}/fpm/php.ini
+elif [ -f /usr/local/php/${PHPVER}/etc/php.ini ]; then
+    configure_php_ini /usr/local/php/${PHPVER}/etc/php.ini
+fi
+
+# use a nice name for the php-fpm service, so that it does not depend on php version running. Try to make that work
+# both for docker and VMs
+if [ -f "/etc/init.d/php${PHPVER}-fpm" ]; then
+    ln -s "/etc/init.d/php${PHPVER}-fpm" /etc/init.d/php-fpm
+fi
+if [ -f "/lib/systemd/system/php${PHPVER}-fpm.service" ]; then
+    ln -s "/lib/systemd/system/php${PHPVER}-fpm.service" /lib/systemd/system/php-fpm.service
+    if [ ! -f /.dockerenv ]; then
+        systemctl daemon-reload
+    fi
+fi
 
 # @todo shall we configure php-fpm?
 
 service php-fpm start
 
-# configure apache
-a2enconf php${PHPVER}-fpm
-service apache2 restart
+# reconfigure apache (if installed). Sadly, php will switch on mod-php and mpm_prefork at install time...
+if [ -n "$(dpkg --list | grep apache)" ]; then
+    echo "Reconfiguring Apache..."
+    if [ -n "$(ls /etc/apache2/mods-enabled/php* 2>/dev/null)" ]; then
+        rm /etc/apache2/mods-enabled/php*
+    fi
+    a2dismod mpm_prefork
+    a2enmod mpm_event
+    a2enconf php${PHPVER}-fpm
+    service apache2 restart
+fi
+
+echo "Done installing PHP"
